+++ CloudShell.Providers.Applications/ApplicationProviderExtension.cs
@@ -75,6 +75,7 @@ public sealed class ApplicationProviderExtension : ICloudShellExtension
         builder.Services.TryAddSingleton<ContainerApplicationOrchestratorServiceDescriptionOperations>();
         builder.Services.TryAddSingleton<ApplicationContainerOrchestratorServicePreparationOperations>();
         builder.Services.TryAddSingleton<ApplicationContainerImageMaterializer>();
+        builder.Services.TryAddSingleton<ContainerApplicationIngressOperations>();
         builder.Services.TryAddSingleton<ContainerApplicationDeploymentDescriptionOperations>();
         builder.Services.TryAddSingleton<ContainerApplicationDeploymentOutcomeOperations>();
         builder.Services.TryAddSingleton<ApplicationContainerHistoryService>();